requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Security-Too==5.3.0
Flask-Migrate==4.0.4
SQLAlchemy==2.0.19
Werkzeug==2.3.7
openpyxl==3.1.2
python-dotenv==1.0.0
email-validator==2.0.0